Create README.md
Machine learning is a method of data analysis that automates analytical model building. It is a
branch of artificial intelligence based on the idea that systems can learn from data, identify patterns
and make decisions with minimal human intervention. This is now widely used in across all the
domains, starting from classification problems to humanoid robot.